requirementsMAC_M1.txt
absl-py==1.4.0
altgraph==0.17.3
async-generator==1.10
attrs==22.2.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dill==0.3.6
docutils==0.19
exceptiongroup==1.1.1
flatbuffers==23.3.3
fonttools==4.39.3
idna==3.4
Kivy==2.1.0
Kivy-Garden==0.1.5
kiwisolver==1.4.4
macholib==1.16.2
matplotlib==3.7.1
mediapipe-silicon==0.9.1
mypy-extensions==1.0.0
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
outcome==1.2.0
packaging==23.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.2.0
protobuf==3.20.3
pydantic==1.10.7
Pygments==2.14.0
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
pyparsing==3.0.9
PySnooper==1.1.1
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
typing_extensions==4.5.0
urllib3==1.26.15
